Introduce separate camera actor
This change introduces a new actor that will be used to control
the game's camera. This funtionality needed to be moved out of
the PlayerCharacter in order to properly support local coop.
Currently, this contains a placeholder implementation that focuses on
the default character (controller idx 0). Later, it can be used to keep
both players properly in the frame.