Enable keyboard input forwarding to multiple players
This is done by introducing a custom ViewportClient which forwards
key input coming from the keyboard to all players. This is needed since
UE sends the Keyboard input only to the local player0, which makes
sense for networked multiplayer.
Currently both players are controlled via the same input. The next
step is to introduce separate keyboard bindings for players.