Control camera by tracking player controllers
The BmCameraActor now uses a set of "tracked" PlayerControllers to
adjust both its 2D position as well as its height. This leads to the
camera both moving as well as zooming to always keep the players in
frame. Additionally, a minimum and maximum height can be set to control
how far in/out the camera is allowed to zoom at maximum.