Control game setup with BmGameMode
BmGameMode now serves a few additional purposes.

It has a set of BmInputBindings which can be set via the Editor. This
InputBindings are distributed to the PlayerControllers based on their
ControllerID when their Pawn is spawned.

CreatePlayer can be used to set up a player by providing which
ControllerID it should use. If a controller for that ID does not exist,
it will be created. Additionally, a pawn will be spawned for each player
created. For ID 0, the default pawn spawned by the game is used.

BmGameMode also "registers" players by adding them to the tracked
controllers of BmCameraActor.