Implement Base Bomb Actor
BmBaseBombActor will serve as the base class for other bomb types.
It implements the basic functionality of exploding and finding hit
actors in C++ (via line tracing the world).

The class' functionality is exposed to blueprints so that future bomb
types can easily be implemented in Blueprints.

Bombs trigger a delegate when exploding in order to notify interested
actors before they are destroyed. This can be used by the
PlayerCharacter to keep track of the placed bombs.