Implement bomb placing functionality
PlayerCharacters now have the ability to place bombs. Once triggered,
a new actor of the bomb type is spawned. The bomb type can be changed
dynamically via a property in the character to allow for pick-up based
bomb types.
Players are only allowed to place a maximum number of bombs
at once, defined by maxBombCount.

For simplicity, triggering the bomb place action also broadcasts a
delegate using the last placed bomb as its argument. This can be used
to implement remote bombs.