Implement Base Pickup
BmBasePickupActor, similarly to how bombs work, will serve as the base
for other Pickup types, which may also be implemented in Blueprints.

The base class sets up common functionality, such as the collision
response and performs a test to check whether the overlapped actor
is a PlayerCharacter. If yes, OnCharacterPickup() is called to, which
can be overriden to create custom Pickup behaviour.