Implement increased bomb range pickup
This pickup type is implemented as a Blueprint class that sets an
override of the range of bombs. Basic bombs have 500 range, the pickup
sets this to 1000.

The override is stored in the PlayerCharacter and starts at 0. If the
override is set to a value > 0, this value is set on a bomb during
spawning.